+++ CHANGELOG.md
@@ -9,3 +9,12 @@
 ## 0.0.3
 
 * Update secondary dependencies.
+
+## 0.0.4
+
+* Added new parameters.
+1. Border color
+2. Color in brightness
+3. Duration of transition
+4. Transition offset
+5. Text color